Update patch set 1
Patch Set 1:

Thanks. +1 for talking about this tomorrow. In the meanwhile, one more thing:

> Being able to support ILA in constraints is mandatory in order to support parallel links

I agree with this use case. Is this possible with the current develop? I quickly went through the code, but it isn't obvious to me.

BTW, it also looks to me that only including ILAs is not enough for this use case. Consider a scenario of two parallel links between ROADMs that are "close" to each other so that there are not ILAs in between. I think that the (physical) fiber paths might well be disjoint so it makes sense to be able to pick and choose arbitrary "GNPy node by its UID" for passing through.

What I'm advocating is request.py only using these "needs-to-go-through" constraints as-is, with no additional postprocessing.

Patch-set: 1
Reviewer: Gerrit User 1009907 <1009907@d5d70762-12d0-45a1-890d-524b12d3f735>