CI: generate a diff of coverage
TODO: move all of the Ansible stuff to oopt-zuul-jobs

Generate a report of how code coverage changes by the proposed code
change. This requires building and testing the project twice, once "as
usually", perhaps via a `tox-cover` jobs, and then once more, but using
the previous version of the source code. Both jobs should generate XML
reports about code coverage and push them to Zuul as artifacts. A third
job then generates an aggregate diff report.

The third job needs access to two separate checkouts of the source code,
once for the previous version (which is conveniently provided by Zuul as
`origin/{{ zuul.branch }}`, and then for the new state (as provided by
default in Zuul). Multi-project jobs are not fully supported because
this patch only provides "this project"'s source tree to pycobertura. I
think that anything more involved will require extra commands to
pycobertura to inform it about other locations of source from all extra
projects. When building projects, however, the previous state of *all*
repositories, not just of "this project", is correctly checked out.

Note that the "previous state" includes everything except Zuul's own
configuration. This includes any auxiliary shell scripts and/or tool
config files (tox, pytest,...) that might affect how the build/test
runs. This also means that when adding support for coverage diffing to a
project, one has to first make sure that coverage *reporting* works
first, and only then configure a project to use these jobs for actual
coverage diffing.

Here's how a simple Python project (based on the usual zuul-jobs' tox
jobs) might be extended to add coverage diffing:

  +- job:
  +    name: myproj-tox-cover
  +    parent: tox-cover
  +    post-run: playbooks/coverage-diff/upload.yaml
  +    vars:
  +      coverage_xml_dir: "{{ zuul.project.src_dir }}/cover"
  +
  +- job:
  +    name: myproj-tox-cover-previous
  +    parent: myproj-tox-cover
  +    pre-run: playbooks/coverage-diff/checkout-previous.yaml
  +    vars:
  +      coverage_artifact_name: "Coverage data (previous state)"
  +
  +- job:
  +    name: coverage-diff
  +    parent: base
  +    description: |
  +      Report how coverage changed compared to the previous state of the repo.
  +    pre-run: playbooks/coverage-diff/prepare.yaml
  +    run: playbooks/coverage-diff/generate.yaml
  +    success-url: coverage-diff.html
  +    failure-url: coverage-diff.html
  +    vars:
  +      previous_source_prefix: "{{ ansible_user_dir }}/previous-src"
  +      coverage_xml_dir: "{{ zuul.project.src_dir }}/cover"
  +      coverage_job_name_previous: myproj-tox-cover-previous
  +      coverage_job_name_current: myproj-tox-cover
  +
   - project:
       check:
         jobs:
  -        - tox-cover
  +        - myproj-tox-cover:
  +            provides: coverage-new
  +        - coverage-diff:
  +            requires:
  +              - coverage-old
  +              - coverage-new
  +            dependencies:
  +              - myproj-tox-cover-previous
  +              - myproj-tox-cover
  +        - myproj-tox-cover-previous:
  +            provides: coverage-old
  +            voting: false
           - tox-docs

Here, `tox-cover` is replaced by `myproj-tox-cover` which is configured
to publish one new artifact, the coverage report in the Cobertura XML
format. A second job is then added, deriving from the first one, but set
up to use the original state of git repos as before the change that is
currently being tested. Finally, a last job takes artifacts from jobs #1
and #2, and runs `pycobertura` on them. Its exit status can be used for
checking whether the patch introduces untested code (which is probably a
good candidate to a non-voting status as well in real-world projects).

Change-Id: I6847139620419fc8f712b120455305c98d27df41