CI: run pyflake8 for coding style nagging in changes
This combines a few pieces of magic:

- Zuul merges speculatively, but in the git trees that the build VMs
see, the "current branch" is always the state as-if the currently tested
change was merged, while origin/"current branch" is the result of all
previous changes, if any, applied to the current tip of the target
branch.

- flake8 has a --diff option which appears to work correctly, *but* it
reports on all lines that are present in its input, including the
context -> hence the -U0 option.

Change-Id: I28403ff0132fac0b52696c82306732a4f81fa66a
Depends-on: https://review.gerrithub.io/c/Telecominfraproject/oopt-zuul-jobs/+/490839
Seel-also: https://gitlab.com/pycqa/flake8/issues/58