Find paths with include nodes constraint
The compute_constrained_path function in request.py handles include nodes constraint by first finding all simple paths shorter
then a cutoff length and then checking if any of the paths meet the constraint. The cutoff is currently set to 120 which is
too small for large networks. For the CORONET_CONUS_Topology.json topology, for example, the shortest path between some node
pairs is longer than the cutoff.

This is a small proposed change to first compute the length of the shortest path and then set the cutoff to at least 20% (a
conservative number) higher than this. I also propose to increase the minimum cutoff from 120 to 150.

Change-Id: I97ff2915fb38e4681bd64d60f2cd6dfd422afd4f
Signed-off-by: Jonas Mårtensson <jonas.martensson@ri.se>