Update patch set 7
Patch Set 7:

(1 comment)

> I think I have understood and agree:
> first you'll merge this commit and then 489922, am I right ?

Yes, this is how it works in Gerrit. If this patch has a parent (as in "git parent commit"), then this patch can only be merged after (or at the same time) as the parent. Unless one explicitly removes this relation via a rebase on top of something else.

Patch-set: 7