linters: Increase maximum line length
The code is already using long lines, so I think we can avoid a
discussion whether to raise this limit, and instead we can focus on
another discussion :) on what that limit should be.

Random googling suggests that GitHub's code renderer operates at 119
chars per line. A random style guide mentions going to 100 or 120. Prior
to this change, there has been 712 violations of PEP8 check E501.
Picking 119 would bring us down to 21 violations, and using 120 reduces
it further to 16. I think that 120 is a cuter number than 119, so 120 it
is unless someone objects hard enough to propose another cutoff in a
follow-up patch.

Change-Id: I57f48fcb6846fea35223daac91aa2e8c7afabc63