linters: Increase maximum line length
The code is already using long lines, so I think we can avoid a
discussion whether to raise this limit, and instead we can focus on
another discussion :) on what that limit should be.

Random googgling suggests that GitHub's code renderer operates at 119
chars per line, so I think it's a reasonable limit. Doing this brings us
from 712 violations of E501 down to 21.

Change-Id: I57f48fcb6846fea35223daac91aa2e8c7afabc63