Update patch set 1
Patch Set 1:

> Sorry, but in my editor if I press return while in an indented block, whitespace corresponding to indentation is added to the blank line. Having to change that manually every time would be inconvenient.

Mine behaves the same, but when I save and/or go to the next line, the whitespace gets deleted ("prune empty lines" I think). This coding style requirement comes from PEP8, so it is not GNPy-specific.

I'll amend this commit in the meanwhile, then. I'm not sure what your editor is, but I'll be happy to help set it up and/or add a git commit hook which takes care of this if you'd like to.

Patch-set: 1