Update patch set 5
Patch Set 5:

> It seems that since commit c41cddfff543e46b29f2496949d82f12eb1451a4 (by Andrea D'Amico, merged this March), the code assumes that for each instance of any subclass of the `Parameters` class, all "properties" (i.e., anything marked with the @property decorator) can be copied/serialized/passed-to-the-constructor via a manual detour over another member with the same name, but prefixed with an underscore. I don't know why it was coded that way, so better please have Andrea comment here. (It's one more reason to avoid using two properties for one purpose as I outline in the JSON comment below, IMHO.)
> 
> I'm also not happy that the test suite run now takes 4:30 min instead of 1:20 min it used to take before this patch.


The method 'asdict()' was defined in order to reproduce the previous implementation that used the nametuple. The purpose is to obtain a dictionary object from an instance of a 'Parameters' class that is equivalent to the dictionary(uploaded from a configuration json file) used to initialize the instance itself. An alternative to the implemented solution would be to store the input dictionary to the constructor in a dedicated attribute and directly return it as dictionary. I have decided to exclude this solution to avoid any dependency of the instance on the original configuration json file, because the attribute values of a 'Parameters' class can be different from the values expressed in the original configuration json file (to be specific, the length attribute in the FiberParams is converted in meters). The error that arises in the test is due to the fact that the 'asdict' method requires that the property as a corresponding private attribute defined in the class. To overcome the issue, it would be sufficient to include the definition of the attribute 'self._lumped_losses' in the constructor.

Patch-set: 5
Reviewer: Gerrit User 1017876 <1017876@d5d70762-12d0-45a1-890d-524b12d3f735>