Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> (1 comment)
> 
> > Patch Set 5:
> > 
> > (1 comment)

Why would we need to have separate elements for ingress and egress? Couldn't we just have one element per degree (and maybe per add/drop block as well but based on the ROADM discussion in the last call this will probably be needed anyway)? I was not involved in the demo so I don't know exactly what the problem was there.

Patch-set: 5