Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > (1 comment)
> 
> Ok, but I guess in this case I still don't really see the point. I mean the new derived class literally just calls super().__init__ (which, BTW, I think is unnecessary since the init function of the base class will be called anyway if it is not overridden).

Yes, it is like you are saying. The only reason why at the calls we have decided this solution is to divide in more in more steps the introduction of this new feature. Doing in such way, the new code is just added to the original one without modifying it.

Patch-set: 5