Update patch set 5
Patch Set 5:

> > Even with this, it was still a problem for booster and preamp in the line card because both were of a different type, yet for ONOS they were hidden in just one box (along with the WSSes that also affect their functionality). So our "Line Degree ROADM 1U pizza box" was modeled by three GNPy elements.
> 
> This is where I don't follow. I don't understand what the problem is or why you need 3 GNPy elements to model one degree.

That degree contains a booster EDFA and a preamp EDFA. Each of these EDFAs must be a GNPy element. I agree that this is not really a GNPy problem, but rather a problem of topology mapping. This should be solvable with that YANG work with IETF-TE models for topology mapping, hopefully.

> > A much bigger problem was "directionality" of GNPy's ROADM element. All other GNPy elements but ROADM simply propagate spectrum from input to output, but a GNPy ROADM has multiple possible inputs and outputs. Even though our (physical) ROADMs use the same EDFA model in Line Degrees and Add Drops, we still needed different power targets as shown here
> 
> Yes, but why is it not sufficient to have one GNPy element per degree to model different power targets?

I'm still assuming that we keep our internal GNPy nomenclature where any amplifier (apart those in the A/D SRG...) is modeled as a separate element. I.e., there is no amplifier in this GNPy ROADM, they are all external and explicitly modeled. Then the topo looks roughly like this:

         +---------+                    (P2)                    +---------+
    ---->| preamp  |-->\    +--------+   |    +--------+    /-->| booster |---->
         +---------+    \   | ROADM  |   v    | ROADM  |   /    +---------+
                  (P1)-> *--+ Degree +--------+ Degree +--*
         +---------+    /   |   X    |        |   Y    |   \    +---------+
    <----| booster |<--/    +-----+--+        +-+------+    \<--| preamp  |<----
         +---------+               \            /               +---------+
                             (P3)-> \          /
                                     \        /
                                    +-+------+-+
                                    |   ROADM  |
                                    | Add/Drop |
                                    +----------+

ROADM "Degree X" can be crossed either:

- P1 -> P2,
- P2 -> P1,
- P1 -> P3,
- P3 -> P1.

For the demo I wanted P_ch(P2) = -12dBm, P_ch(P1) = -25dBm, P_ch(P3) = -12dBm. I cannot do that today with GNPy's develop branch because that code really wants P1 = P2 = P3.

(Note that I currently cannot let GNPy know that there's no path between P2 and P3, but that's not relevant for this discussion, that's a separate bug.)

Patch-set: 5