Update patch set 12
Patch Set 12:

> Patch Set 12: Code-Review-1
> 
> (5 comments)
> 
> Thanks for addressing my comments. I would appreciate if that _beta2 polarity fix is done in a newly created commit (as a parent of this one). Please let me know if you require some help with this.
> 
Ok, I will do it. I would appreciate if you can give some more details on how to do it.

> I have no idea if the change is factually correct, but I of course feel comfortable just taking your word for that :).
> 
Yes, actually it was wrong before. We never noticed that because, if all the fibers in field has the same sign - and this is the case for modern WDM networks - it does not change the performance.

> As a followup work, we can extend transponder mode selection to take CD into account (via adding an equipment_config threshold for each mode, similar to how we handle GSNR now).

Yes, I agree and I think that also the DGD upgrade should go in that direction.

Patch-set: 12