Bug fix in converting the dispersion D in beta2
The actual conversion formula includes the minus (-), not the absolute
value. We never noticed it as GNPy simulates only modern networks
which have not DCUs. In this case, the sign of beta2 along a path is
the same for all the spans and, all positive or all negative beta2
don't change the amount of NLI.

Change-Id: I60a61d00c578a1a0436231a2bda8e3b6256fc8b3