tests: call our example entry points via functions
I would like to avoid that extra fork to a child Python interpreter (it
looks like something that can be easily avoided). It's something that's
possible now that the code ships just some trivial wrappers (which are,
in turn, needed for setuptools' `console_scripts`).

This cannot use the `capsysbinary` fixture for wrapping of stdout/stderr
due to something in pytest which already got fixed, but has not been
released yet (May 2020).

Change-Id: Ic4a124a5cbe2bd24c56e4565d27d313fe4da703f
Bug: https://github.com/pytest-dev/pytest/pull/6926