Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > This patch does not work on my local computer (and later commits do not work either with same error)
> > python3 transmission_main_example.py 
> > Traceback (most recent call last):
> >   File "transmission_main_example.py", line 13, in <module>
> >     from gnpy.core.equipment import load_equipment, trx_mode_params
> > ImportError: bad magic number in 'gnpy.core': b'\x03\xf3\r\n'
> > 
> > I do not understand why...
> 
> One major change that this patch is doing is removal of that top-level gnpy/core/__init__.py file. Technically, it's not needed, but it's quite possible that now that it's gone, there's some older gnpy/core.pyc file somewhere in your system, and Python no longer knows that it's obsolete. Chances are that, previously, this gnpy/core/__init__.py was newer than the cached *.pyc, and that is no longer the case.
> 
> What version of Python do you use, and on what platform is that? This looks like some left-over *.pyc from a previous Python version (not sure if 2.x -> 3.x move or something smaller). Have you ever tried to run GNPy with Python 2?
> 
> What should -- I think -- help is removing all of the *.pyc files in:
> 
> - the location where you cloned this repo,
> - in your user Python path (below ~/.local)
> 
> Also please check your system if you happen to have either "gnpy" or "core.pyc" anywhere in the system. If so, that should be cleaned.

I am working with python 3.6 from the beginning on a VM with ubuntu18.04.4 LTS. 
I have removed the .pyc that was on the system, as well as the __pycache__ 
but does not seem to change. There are no .pyc left in the .local (related to gnpy, but there are many related to other packages).

Besides, I though that having a __init.py__ enabled to treat the directories as containing packages... for example callable from other programs. This is a usefull functionality in my opinion, do we really need to remove this file?

Patch-set: 1
Reviewer: Gerrit User 1017751 <1017751@d5d70762-12d0-45a1-890d-524b12d3f735>