Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> thanks , I managed to have it working with the "find . -name \*.pyc -delete" 😊
> However I am stil not sure about deleting the __init__.py with respect to making gnpy a package
> 
> could you please comment on that ?

Summarizing what I said on the call -- it seems that __init__.py presence does not make a difference; regardless if they are present or not, one can always do `import gnpy`, or `import gnpy.core`, etc. What is relevant are documentation. If there's an __init__.py and if it has a docstring, then that thing is used in Python REPL's `help()` built-in. I implemented these in https://review.gerrithub.io/c/Telecominfraproject/oopt-gnpy/+/494298 .

Patch-set: 1