CI: Run Python 3.6 and 3.8 builds on native packages
Let's try to use a distro with long enough support timeframe and a
native Python 3.6 installation instead of relying on upstream Python and
installing it from scratch every time. That would be CentOS 8.

For Python 3.8, the latest Fedora sounds like path of least resistance.
However, there isn't that much space in tmpfs on the 1GB VMs, so let's
use something bigger (that's, ironically, easier to change than the env
variables...).

Change-Id: Ic19bf58875ba6d6afaee77f5c9b467261ec686d0