Add support for exclude node constraints
This implements support for exclude node contraints using the compute_k_constrained_paths function from the previous commit on this topic. Supporting exclude nodes is actually much simpler than include nodes (just remove the nodes from the graph before path computation) so why not do it. In order to not break backwards compatibility for existing scripts, exclude_nodes_list is added at the end of the RequestParams namedtuple and set to an empty list by default. I don't know if this is the best way to do it but it is kind of difficult with namedtuple to add features without breaking compatibility.

Change-Id: I52c98b33973370129514a90405e722344308343d
Signed-off-by: Jonas Mårtensson <jonas.martensson@ri.se>