CI: Restore functionality via running on  Python 3.6 on native packages
We have declared that we're supporting Python 3.6 as the minimal
required version of Python. Let's enforce this via testing. Let's also
do it via a long-term supported distro (CentOS8 in this case) so that we
can have someone else maintain stuff for us; it's easier to rely on
these prebuilt packages instead of downloading Python from upstream all
the time.

Also, Python 3.7 builds are currently broken in the CI because of some
breakage, either in the upstream zuul-jobs repo, or in Vexxhost's VM
images. So, now that we have Python 3.6 VMs available, use this to
restore CI functionality. That part can be reverted in future when
Debian Buster-based builds are working again.

I've also tried to use Python 3.8 on latest Fedora (version 32). There
were a couple of gotchas, including /tmp on tmpfs with not enough space
for building all required packages. Also, there's another issue with
building Pandas from scratch, so let's defer using Python 3.8 until
later. It's something that I will definitely look into, though, because
of significant speed improvements in that distro.

Change-Id: Ic19bf58875ba6d6afaee77f5c9b467261ec686d0
Depends-on: https://review.gerrithub.io/c/Telecominfraproject/oopt-zuul-jobs/+/494597