Update patch set 4
Patch Set 4:

> Patch Set 1:
> 
> Could you please elaborate on this a bit more? As far as I can see, all the formulas for calculating NLI use the absolute value of beta2 so how could the sign have made a difference? How are DCUs relevant?

It is true, so, from the GN-model perspective the prediction doesn't change. What really change is the true effect of the NLI. The true propagation of a signal through two fiber spans having the same dispersion sign is different with respect to one having opposite signe. Keep in mind that the GN-model has been developed for uncompensated transmission (UT) systems.

Patch-set: 4