Update patch set 2
Patch Set 2:

> The autodesign is a result from GNPy especially in the case of planning. For me it is not an option.
> Maybe I miss some reason why you need to remove both convert and autodesign output.

The entry point is called "transmission main example", and the docs say that this how to run a simulation. I agree that autodesign is an important feature of GNPy, but then perhaps we should have a tool called "gnpy-optimize-edfa" for this and not stick everything into a script which says it's propagating light from point A to point B. What do you think?

Patch-set: 2