Update patch set 1
Patch Set 1:

I've tested the patch on Fedora and Debian. On Fedora and Debian I didn't encountered any problem.
My only concern is about the possibility of using the debug mode of an IDE (I use PyCharm) as, in this case I cannot just execute a .py file. Is it still possible somehow to run the gnpy-transmission-example in debug mode?

Patch-set: 1
CC: Gerrit User 1017851 <1017851@d5d70762-12d0-45a1-890d-524b12d3f735>