CI: provision secrets for uploads to PyPI
Originally I was hoping that this is possible via the parent job
collection (oopt-zuul-jobs), but that's not the case, unfortunately, and
I have to do this in this project.

I created the token at PyPI's own web UI for my account, and saved the
secret to a file. Then I executed Zuul's helper script:

  $ ~/work/prog/openstack/zuul/zuul/tools/encrypt_secret.py \
    --infile secret-pypi --tenant TelecomInfraProject \
    https://zuul.vexxhost.dev/ Telecominfraproject/oopt-gnpy

Change-Id: I05ab853ae26a449212e832e0f7e261c9ed71e364