docs: do not talk about branches anymore
We talked about this during the coders call on 2020-06-02. When this
project started, it was decided to keep the `master` branch essentially
frozen between releases, and concentrate development into the `develop`
branch. Since that time, we've got a decent CI coverage and end-to-end
tests which should guard us against regressions and inaccurate
simulation results. It also seems that our users have a preference to
use tagged releases anyway, so we can get by by deprecating the
`develop` branch.

As agreed during the conf call, from this release on we'll be merging
approved changes directly into the `master` branch, and the `develop`
branch will be retired. We have measures in place that ought to limit
the risk of regression within `master` to an acceptable level.

Change-Id: I0048ba8e36810237667e6930380a9c8570090a26