Update patch set 1
Patch Set 1: Code-Review-1

During the codes call this week we decided to put this into the documentation and not directly into the README file.

Here's an example of how to add content to the docs. Put the .rst (or .md) file into docs/, and add its name to an appropriate place in docs/index.rst to make it appear in the ToC. My recommendation is to also try to leverage the bibtex extension for managing references; I think there's more than one paper to point to already.

  https://review.gerrithub.io/c/Telecominfraproject/oopt-gnpy/+/495862

Patch-set: 1
Reviewer: Gerrit User 1009907 <1009907@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1