Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Let's bring back the comments from GitHub for clarity (yeah, sorry for the platform change):
> 
> > > Is there a reason why Transceiver.__call__ does not simply introduce 
> > > an appropriate amount of ASE noise to each channel of the SpectrumInformation? 
> > > That way, there would be no need to mess with incremental SNR additions.
> > > 
> > > To me, it's rather counter intuitive that it's OK to perform the whole 
> > > simulation/propagation, and just before we're about to print the result, 
> > > we magically adjust the SNR at the start. Surely the SNR "present from 
> > > the beginning" must have had some effect along the path as well?
> > 
> > 
> > @jktjkt, good question, I agree that it's counter-intuitive.
> > 
> > The ASE noise can't be introduced in Transceiver.__call__, since then it
> > would be introduced at both the transmit and receive ends. 
> > 
> > My original proposal (PR #177), which was merged but later reverted, was 
> > to introduce the transmitter ASE noise in create_input_spectral_information. 
> > It turned out that this increased the computation time of the 
> > propagate_and_optimize_mode function, see discussion in PR #186 and 
> > issue #185. I then proposed an optimization in PR #197 but that was rejected 
> > in favor of instead adding the transmitter ASE noise after propagation, 
> > see PR #199.
> 
> A bit of brainstorming -- what about adding this in Transceiver.__call__, but only in case the SNR is at +infinity? That way it would effectively only get added once, at the beginning of the transmission.
> 
> Or is that too hackish in your book?

I'm not sure I understand your proposal correctly but I think you're missing the point (I didn't explain it very well in my previous comment). The update_snr function does not impact the spectral info. It just updates the already calculated SNR. We already do this for the last transceiver even though the actual ASE contribution comes from the first transceiver in the case of tx_osnr (this is the "hackish" part). This patch just proposes a small change to also update the SNR for the first transceiver. If your proposal is to actually add the noise to the carriers of the spectral info it doesn't work since it increases computation time as described before. Or is your proposal is to just move the call to update_snr from request.py to Transceiver.__call__? I thought about that but could not come up with a clean way to identify whether the transceiver is the first or last (this is not really a problem for tx_osnr since that should be added to both first and last but add_drop_osnr should only be added to the last). For example, the OSNR_ASE could still be +inf after propagation in the case where there are no amplifiers in the chain. Anyway, moving it to Transceiver.__call__ would not have any impact on the actual calculation.

Patch-set: 2