Update patch set 2
Patch Set 2: Code-Review+2

> Or is your proposal is to just move the call to update_snr from request.py to Transceiver.__call__? I thought about that but could not come up with a clean way to identify whether the transceiver is the first or last (this is not really a problem for tx_osnr since that should be added to both first and last but add_drop_osnr should only be added to the last).

Yes, that's what I had in mind. I assumed that if the per-carrier SNR passed to __call__ is +infinity, then we know it's a first transponder, otherwise it's the last one. However, you raised a point about the case with no EDFAs, so my proposal won't fly.

It might be interesting to revisit this in future by making the transponders actual *sources* of carriers, but that would be a long shot now.

Patch-set: 2
Reviewer: Gerrit User 1009907 <1009907@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2