Update patch set 3
Patch Set 3: Code-Review-1

(5 comments)

This patch adds an unused function that is not covered by tests, and a significant part of that function is similar to compute_constrained_path().

I think that the path of least resistance is to consider that new method a helper, and call it with k=1 from within compute_constrained_path().

Patch-set: 3
Reviewer: Gerrit User 1009907 <1009907@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1