Raise error when parameters required for Raman simulation are missing
If any of nli_parameters or raman_parameters is missing from the
SimParams input file, a simulation with Raman fibers will crash. I think
it's better to catch it early and raise a ParametersError in this case.
The SimParams are only used by the RamanFiber and there is no point in
passing this input file when Raman fibers are not used so I think it
makes sense to make these parameters required for the SimParams input.

Change-Id: I2446fb81588597f4eb090e418cf01aed5b7c1ae6
Signed-off-by: Jonas Mårtensson <jonas.martensson@ri.se>